* UPD: More fixes build errors and warnings. Minor cleanup. * FIX: Fixed has_home working. (Thx @dsh2dsh)

+++ code/engine.vc2008/xrCore/StringsUtils.h
@@ -19,7 +19,7 @@ const char* ConvertToUTF8(const xr_string& pStr, string64& OutUtf8String)
 {
 	R_ASSERT(pStr.size() < 64);
 	wchar_t utfRusStr[64] = { 0 };
-	MultiByteToWideChar(CP_ACP, 0, pStr.c_str(), pStr.size(), utfRusStr, sizeof(utfRusStr));
+	MultiByteToWideChar(CP_ACP, 0, pStr.c_str(), pStr.size(), utfRusStr, (int)sizeof(utfRusStr));
 
 	WideCharToMultiByte(CP_UTF8, 0, utfRusStr, sizeof(utfRusStr), OutUtf8String, sizeof(OutUtf8String), 0, 0);
 	return &OutUtf8String[0];
